Rename "Bucket#shaders" to "Bucket#shaderInterfaces"
fixes #2172
