+++ test/js/data/line_bucket.test.js
@@ -6,7 +6,7 @@ var path = require('path');
 var Protobuf = require('pbf');
 var VectorTile = require('vector-tile').VectorTile;
 var Point = require('point-geometry');
-var LineBucket = require('../../../js/data/line_bucket');
+var LineBucket = require('../../../js/data/bucket/line_bucket');
 
 // Load a line feature from fixture tile.
 var vt = new VectorTile(new Protobuf(new Uint8Array(fs.readFileSync(path.join(__dirname, '/../../fixtures/mbsv5-6-18-23.vector.pbf')))));
